fix: render Nextra docs Layout behind a Client Component boundary
Production Docker still digested with theme deps present; keep getPageMap on the server and mount Layout/Navbar/Footer from a client shell.
